requirements.txt
google-genai>=1.52.0
Pillow